+++ codeforces_core/__init__.py
@@ -5,5 +5,8 @@ __all__ = (
     "contest_list",
     "contest_register",
     "contest_meta",
+    "contest_standing",
+    "problems",
+    "httphelper",
 )
 __version__ = "0.0.1"