+++ codeforces_core/submit.py
@@ -37,7 +37,7 @@ async def async_submit(http: AioHttpHelperInterface, contest_id: str, level: str
           # http = HttpHelper(token_path='/tmp/cache_token', cookie_jar_path='/tmp/cache_cookie_jar')
           http = HttpHelper(token_path='', cookie_jar_path='')
           await http.open_session()
-          result = await async_login(http=http, handle='<handle>', passwd='<password>')
+          result = await async_login(http=http, handle='<handle>', password='<password>')
           assert(result.success)
 
           print('before submit')